Fix the wrong distance displayed on the TBT UI
Root cause:
LC could receive the unit string which has been translated to
different language. In this case, it breaks the condition where
converting distance to floating point number for TBT UI by
checking "km" / "mi".

Countermeasure:
Remove the unnecessary condition of checking "km"/"mi" and
change distance data type to float then directly assign the received
distance to it after converting from byte to floating point number.
Besides, remove ".0" for received distance.

Testing and Verification Instructions:
  1. Connect BT and Navi app on iPhone.
  2. Start route guidance.
  3. Access NAV03 or NAV05, check if correct distance is displayed
     when distance is larger than 1000m.
  4. Verify step 1 to 3 by using Android.

Issue: Autooem-IXWW22-6309
Type: Bug Fix
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/87170
Tested-by: Liao, Edward <edward.liao@garmin.com>
Approve-Change: Liao, Edward <edward.liao@garmin.com>

+++ navi/NAVI_pub.h
@@ -89,7 +89,7 @@ typedef struct
     {
     uint32_t list_idx;
     uint32_t icon_idx;
-    uint16_t distance;
+    float    distance;
     char     dist_unit[MAX_TBT_DIST_UNIT_SIZE];
     char     description[MAX_TBT_DESCRIPTION_SIZE];
     }navi_tbt_data_type;
@@ -117,7 +117,7 @@ void NAVI_get_alert_distance( char** dist );
 bool NAVI_is_tbt_message_displayed( void );
 uint16_t NAVI_get_tbt_list_size( void );
 void NAVI_reset_tbt_buffer( void );
-bool NAVI_get_tbt_item( const int tbt_index, uint32_t* icon_index, uint16_t* distance, char** dist_unit, char** description );
+bool NAVI_get_tbt_item( const int tbt_index, uint32_t* icon_index, float* distance, char** dist_unit, char** description );
 
 #ifdef __cplusplus
 }
